[Networking] - Move Network controls into a folder.
